RLbDS - Reinforcement Learning based Dynamic Scheduling/requirements.txt
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
pyyaml==6.0
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
tqdm==4.65.0